+++ packages/engine-multi/test/integration.test.ts
@@ -83,7 +83,10 @@ test.serial('trigger job-complete', (t) => {
 
     api.execute(plan).on('job-complete', (evt) => {
       t.deepEqual(evt.next, []);
-      t.true(evt.duration < 20);
+      t.log('duration:', evt.duration);
+      // Very lenient duration test - this often comes in around 200ms in CI
+      // (because of parelleisation in the tests I think)
+      t.true(evt.duration < 200);
       t.is(evt.jobId, 'j1');
       t.deepEqual(evt.state, { data: {} });
       t.pass('job completed');