+++ packages/ws-worker/test/util/worker-token.test.ts
@@ -2,9 +2,12 @@ import test from 'ava';
 import * as jose from 'jose';
 
 import generateWorkerToken from '../../src/util/worker-token';
+import { createMockLogger } from '@openfn/logger';
+
+const logger = createMockLogger();
 
 test('should generate a worker token as a JWT', async (t) => {
-  const jwt = await generateWorkerToken('abc', 'x');
+  const jwt = await generateWorkerToken('abc', 'x', logger);
 
   t.truthy(jwt);
   t.assert(typeof jwt === 'string');
@@ -17,7 +20,7 @@ test('should verify the signature', async (t) => {
   const secret = 'abc';
   const encodedSecret = new TextEncoder().encode(secret);
 
-  const jwt = await generateWorkerToken(secret, 'x');
+  const jwt = await generateWorkerToken(secret, 'x', logger);
 
   const { payload, protectedHeader } = await jose.jwtVerify(jwt, encodedSecret);
 
@@ -29,13 +32,13 @@ test('should throw on verify if the signature is wrong', async (t) => {
   const secret = 'abc';
   const encodedSecret = new TextEncoder().encode('xyz');
 
-  const jwt = await generateWorkerToken(secret, 'x');
+  const jwt = await generateWorkerToken(secret, 'x', logger);
 
   await t.throwsAsync(() => jose.jwtVerify(jwt, encodedSecret));
 });
 
 test('should include the server id in the payload', async (t) => {
-  const jwt = await generateWorkerToken('abc', 'x');
+  const jwt = await generateWorkerToken('abc', 'x', logger);
   const claims = jose.decodeJwt(jwt);
 
   t.is(claims.worker_id, 'x');